Logger value field (#10)
## Description

add value field to logger so we can track a more comprehensive log and
use it in a future feature to reconstruct the dp array

note that this PR does not account for slice operations and needs a
follow-up PR to complete the functionality

drive by:
- moved some tests

## TODO

<!-- Notable points that this PR has either accomplished or will
accomplish. -->

- [x] Logger: add value field to WRITE

## TESTING

- [x] Updated existing tests 

## Questions

<!-- Any concerns or points of confusion? -->

- [ ] @itsdawei i moved some tests- are you fine with this? I just don't
like the idea of importing DPArray into the logger_test, and want
separate files for tests for DPArray, logger, and tests concerning both.
Or we could put the entire dp_array_logger_test into dp_array_test. lmk

## Status

- [x] I have read the guidelines in
[CONTRIBUTING.md](https://github.com/itsdawei/dynamically_programmed/blob/master/CONTRIBUTING.md)
- [x] I have formatted my code using `yapf`
- [x] I have linted my code with `pylint`
- [x] I have tested my code by running `pytest`
- [x] I have added a one-line description of my change to the changelog
in `HISTORY.md`
- [x] This PR is ready to go